% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_zones.R
\name{aggregate_zones}
\alias{aggregate_zones}
\title{Aggregate data between zones
Provides the ability to take attributes from one scale of polygon and split aggregate them to another.}
\usage{
aggregate_zones(
  data,
  geom,
  crosswalk,
  ID = "divide_id",
  fun = "mean",
  join = TRUE,
  drop = NULL
)
}
\arguments{
\item{data}{tabular data reference to the smaller unit}

\item{geom}{sf, sfc, SpatialPolygonsDataFrame, or SpatialPolygons object with polygonal geometries}

\item{crosswalk}{a crosswalk of smaller unit to major unit relations}

\item{ID}{the grouping ID}

\item{fun}{an optional function or character vector, as described below}

\item{join}{if TRUE the geom will be joined to the results}

\item{drop}{colnames to drop from table}
}
\value{
data.frame or sf object
}
\description{
Aggregate data between zones
Provides the ability to take attributes from one scale of polygon and split aggregate them to another.
}
\details{
the cross walk table must have at least 4 columns containing the ID and areas of the small units and
and the ID and areas of the large unit. The name of the large unit IDs is provided by the \code{ID} parameter in the function
signature. The smaller unit IDs must have a corresponding attribute in the data input. The area of the larger units must be called
\code{areasqkm}, while the area of the smaller units must be called \code{s_areasqkm.}
}
