% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{binned_json}
\alias{binned_json}
\title{Geometric Mean Summary}
\usage{
binned_json(x = NULL, coverage_fraction = NULL, bins = 10)
}
\arguments{
\item{x}{vector of values}

\item{coverage_fraction}{coverage fraction}

\item{bins}{number of bins to compute (default 10)}
}
\value{
data.frame
}
\description{
Geometric Mean Summary
}
