% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{circular_mean}
\alias{circular_mean}
\title{Geometric Mean Summary}
\usage{
circular_mean(x, coverage_fraction)
}
\arguments{
\item{x}{vector of values}

\item{coverage_fraction}{coverage fraction}
}
\value{
data.frame
}
\description{
Geometric Mean Summary
}
