% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{.read_nc}
\alias{.read_nc}
\title{Data Extraction from NetCDF}
\usage{
.read_nc(file, dims, cell)
}
\arguments{
\item{file}{a netcdf file path}

\item{dims}{the row and column IDs to extract data between (xmin xmax ymin ymax)}
}
\value{
data.table
}
\description{
Data Extraction from NetCDF
}
