% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{.read_raster}
\alias{.read_raster}
\title{Data Extraction from tif(s) & Raster/spatRaster}
\usage{
.read_raster(file, w, layer = -1)
}
\arguments{
\item{file}{a in-memory Raster(layer, stack, brick), SpatRaster, or file path(s)}

\item{w}{a weight grid}

\item{layer}{the layer to extract from. If -1 (default), get all layers}
}
\value{
data.table
}
\description{
Data Extraction from tif(s) & Raster/spatRaster
}
