% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{.zonal_io}
\alias{.zonal_io}
\title{Data Extraction}
\usage{
.zonal_io(file, w, layer)
}
\arguments{
\item{file}{a netcdf file path, set if tif file paths, or
in-memory Raster(layer, stack, brick), SpatRaster, or file path(s)}

\item{w}{a weight grid}

\item{layer}{the layer to extract from. If -1 (default), get all layers}
}
\value{
data.table
}
\description{
Data Extraction
}
