% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{equal_population_distribution}
\alias{equal_population_distribution}
\title{Equal Area Distribution}
\usage{
equal_population_distribution(value, coverage_fraction, groups = 4)
}
\arguments{
\item{value}{vector of value}

\item{coverage_fraction}{coverage fraction}

\item{groups}{number of intervals to create}
}
\value{
data.frame
}
\description{
Equal Area Distribution
}
