% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_zonal}
\alias{execute_zonal}
\title{Execute Spatial Intersection}
\usage{
execute_zonal(file = NULL, geom = NULL, ID = NULL, w = NULL)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{w}{a precomputed weighting grid produced with \code{weighting_grid}}
}
\value{
data.table
}
\description{
Execute Spatial Intersection
}
