% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_cat.R
\name{execute_zonal_cat}
\alias{execute_zonal_cat}
\title{Execute Spatial Intersection on Categorical Data}
\usage{
execute_zonal_cat(
  file = NULL,
  geom = NULL,
  ID = NULL,
  w = NULL,
  rcl = NULL,
  join = TRUE
)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{w}{a precomputed weighting grid produced with \code{weighting_grid}}

\item{rcl}{a data.frame with columns names \code{from} and \code{to}.
\code{From} should be the categorical values used in the raster,
while \code{to} should be the categorical names to use in the output table headings.
If left NULL (default) the headings will not be altered.}

\item{join}{should output be joined to input geom by the supplied ID (geom cannot be NULL)}
}
\value{
a data.table
}
\description{
Execute Spatial Intersection on Categorical Data
}
