% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{gm_mean}
\alias{gm_mean}
\title{Get geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{na.rm}{Should NA values be removed?}
}
\value{
mode value
}
\description{
Vectorized and suited to handle NAs and negative values
}
