% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{prep_geom}
\alias{prep_geom}
\title{Prep Incoming Data}
\usage{
prep_geom(geom, crs = NULL)
}
\arguments{
\item{geom}{sf or SpatVector object}

\item{crs}{coordinate reference system}
}
\value{
SpatVector
}
\description{
Prep Incoming Data
}
