% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{prep_input}
\alias{prep_input}
\title{Prep Incoming Data}
\usage{
prep_input(data, subds = NULL)
}
\arguments{
\item{data}{SpatRaster, file path, raster}

\item{subds}{subdatasets}
}
\value{
SpatRaster
}
\description{
Prep Incoming Data
}
