% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize Column Inclusions}
\usage{
sanitize(exe, drop = NULL)
}
\arguments{
\item{exe}{executed zonal product}

\item{drop}{colnames to drop from table}
}
\value{
data.table
}
\description{
Sanitize Column Inclusions
}
