% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{ts_extract}
\alias{ts_extract}
\title{Extract and Transform Time Series Data}
\usage{
ts_extract(output, ID)
}
\arguments{
\item{output}{A data frame containing the time series data in wide format. It must have a column with the name matching \code{ID} and other columns
with names in the format "description_Date" (e.g., "temp_2024-01-01").}

\item{ID}{A character string specifying the column name that uniquely identifies rows in the dataset.}
}
\value{
A transformed data frame with the following changes:
\describe{
\item{description}{The variable names extracted from the original column names.}
\item{Date}{A \code{Date} column converted from the original column names.}
\item{value}{The associated values for each combination of \code{ID}, \code{description}, and \code{Date}.}
}
}
\description{
This function processes an input dataset by reshaping it from wide to long format, separating columns into descriptions and dates,
and renaming the resulting dataset for further analysis.
}
