% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weight_grid}
\alias{weight_grid}
\title{Build Weighting Grid}
\usage{
weight_grid(data, geom, ID, progress = TRUE)
}
\arguments{
\item{data}{file path or}

\item{geom}{sf, sfc, SpatialPolygonsDataFrame, or SpatialPolygons object with polygonal geometries}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{progress}{if TRUE, display a progress bar during processing}
}
\value{
a list(data.table, vector)
}
\description{
Returns a data.table with columns for ID, grid_id, X, Y and weight. By default this object is
sorted on the grid_id
}
