% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weight_grid_to_data}
\alias{weight_grid_to_data}
\title{Use Weight Grid to extract SpatRast Data}
\usage{
weight_grid_to_data(data, w, subds = NULL)
}
\arguments{
\item{data}{SpatRaster or file path}

\item{w}{weight grid}

\item{subds}{subdatasets to extract}
}
\value{
data.table
}
\description{
Returns a data.table with raster data attached to weight grid
}
