% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_grid.R
\name{weighting_grid}
\alias{weighting_grid}
\title{Build Weighting Grid
Returns a data.table with columns for ID, grid_id, X, Y and weight. By default this object is
sorted on the grid_id}
\usage{
weighting_grid(file, geom, ID, progress = FALSE)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{progress}{show progress bar for intersection (default = TRUE)}
}
\value{
a data.tabel
}
\description{
Build Weighting Grid
Returns a data.table with columns for ID, grid_id, X, Y and weight. By default this object is
sorted on the grid_id
}
