% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactextract.R
\name{zone_by_ee}
\alias{zone_by_ee}
\title{Execute Zonal Stats by ExactExtract}
\usage{
zone_by_ee(
  data,
  geom,
  ID,
  fun = "mean",
  subds = 0,
  progress = FALSE,
  join = TRUE,
  ...
)
}
\arguments{
\item{data}{SpatRaster or file path}

\item{geom}{summary units}

\item{ID}{the grouping ID}

\item{fun}{summarization function}

\item{subds}{subdatasets to extract}

\item{progress}{if TRUE, display a progress bar during processing}

\item{join}{should data be joined to geom?}

\item{...}{additional arguments passed on to exact_extract}
}
\value{
data.table
}
\description{
execute exactextract more flexiably
}
