% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{zone_by_weights}
\alias{zone_by_weights}
\title{Execute Zonal Stats by Weight Grid}
\usage{
zone_by_weights(data, w, ID, fun = "mean", subds = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{SpatRaster or file path}

\item{w}{weight grid}

\item{ID}{the grouping ID}

\item{fun}{summarization function}

\item{subds}{subdatasets to extract}

\item{na.rm}{should NA values be removed?}
}
\value{
data.table
}
\description{
execute
}
