% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{.read_nc}
\alias{.read_nc}
\title{Data Extraction from NetCDF}
\usage{
.read_nc(file, w)
}
\arguments{
\item{file}{a netcdf file path}

\item{w}{a weight grid}
}
\value{
data.table
}
\description{
Data Extraction from NetCDF
}
