% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{equal_population_distribution}
\alias{equal_population_distribution}
\title{Equal Area Distribution}
\usage{
equal_population_distribution(values, coverage_fraction, n = 4)
}
\arguments{
\item{values}{vector of values}

\item{coverage_fraction}{coverage fraction}

\item{n}{number of intervals to create}
}
\value{
data.frame
}
\description{
Equal Area Distribution
}
