% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_zonal}
\alias{execute_zonal}
\title{Execute Spatial Intersection}
\usage{
execute_zonal(file, w)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{w}{a weighting grid produced with \code{weighting_grid}}
}
\value{
data.table
}
\description{
Execute Spatial Intersection
}
