% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_zonal}
\alias{execute_zonal}
\title{Execute Spatial Intersection}
\usage{
execute_zonal(
  file = NULL,
  geom = NULL,
  ID = NULL,
  FUN = "mean",
  w = NULL,
  join = TRUE
)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{FUN}{an optional function or character vector, as described below}

\item{w}{a precomputed weighting grid produced with \code{weighting_grid},}

\item{join}{should output be joined to input geom by the supplied ID (geom cannot be NULL)}
}
\value{
data.table
}
\description{
Execute Spatial Intersection
}
