% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{execute_zonal}
\alias{execute_zonal}
\title{Execute Zonal Stats}
\usage{
execute_zonal(
  data = NULL,
  geom = NULL,
  w = NULL,
  ID = NULL,
  fun = "mean",
  subds = NULL,
  na.rm = TRUE,
  progress = FALSE,
  join = TRUE,
  drop = NULL,
  ...
)
}
\arguments{
\item{data}{SpatRaster or file path}

\item{geom}{sf, sfc, SpatialPolygonsDataFrame, or SpatialPolygons object with polygonal geometries}

\item{w}{a weight grid (produced with weight_grid)}

\item{ID}{the grouping ID}

\item{fun}{an optional function or character vector, as described below}

\item{subds}{character or boolean to select a sub-dataset. If NULL all are selected}

\item{na.rm}{remove na values?}

\item{progress}{if TRUE, display a progress bar during processing}

\item{join}{if TRUE the AOI will be joined to the results}

\item{drop}{colnames to drop from table}

\item{...}{optional arguments to pass fun}
}
\value{
sf object or data.table
}
\description{
Execute Zonal Stats
}
