% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_zonal}
\alias{execute_zonal}
\title{Execute Spatial Intersection}
\usage{
execute_zonal(
  file = NULL,
  geom = NULL,
  ID = NULL,
  FUN = "mean",
  w = NULL,
  rcl = NULL,
  join = TRUE,
  area_weight = TRUE
)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}

\item{FUN}{an optional function or character vector, as described in details}

\item{w}{a precomputed weighting grid produced with \code{weighting_grid},}

\item{rcl}{a data.frame with columns names \code{from} and \code{to}.
\code{From} should be the categorical values used in the raster,
while \code{to} should be the categorical names to use in the output table headings.
If left NULL (default) the headings will not be altered.}

\item{join}{should output be joined to input geom by the supplied ID (geom cannot be NULL)}
}
\value{
data.table or sf object
}
\description{
Execute Spatial Intersection
}
\details{
The are current 6 built in summary functions:
\itemize{
\item{"mean"}{The mean value of the coverage data}
\item{"mode"}{The most frequent of the coverage values}
\item{"max"}{The maximum of the coverage values}
\item{"min"}{The minimum of the coverage values}
\item{"gm_mean"}{The geometric mean of the coverage values}
\item{"percent"}{The percentage of each covered value (categorical data)}
}
}
