% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_cat.R
\name{execute_zonal_cat}
\alias{execute_zonal_cat}
\title{Execute Spatial Intersection on Categorical Data}
\usage{
execute_zonal_cat(file, w, rcl = NULL)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{w}{a weighting grid produced with \code{weighting_grid}}

\item{rcl}{a data.frame with columns names \code{from} and \code{to}.
\code{From} should be the categorical values used in the raster,
while \code{to} should be the categorical names to use in the output table headings.
If left NULL (default) the headings will not be altered.}
}
\value{
a data.table
}
\description{
Execute Spatial Intersection on Categorical Data
}
