% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric Mean Summary}
\usage{
geometric_mean(x, coverage_fraction)
}
\arguments{
\item{coverage_fraction}{coverage fraction}

\item{value}{vector of value}
}
\value{
data.frame
}
\description{
Geometric Mean Summary
}
