% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{mode}
\alias{mode}
\title{Mode Summary}
\usage{
mode(x, coverage_fraction)
}
\arguments{
\item{coverage_fraction}{coverage fraction}

\item{value}{vector of value}
}
\value{
data.frame
}
\description{
Mode Summary
}
