% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{prep_geom}
\alias{prep_geom}
\title{Prep Incoming Data}
\usage{
prep_geom(geom, crs = NULL)
}
\arguments{
\item{data}{SpatRaster, file path, raster}

\item{subds}{subdatasets}
}
\value{
SpatRaster
}
\description{
Prep Incoming Data
}
