% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal.R
\name{prep_input}
\alias{prep_input}
\title{Prep Incoming Data}
\usage{
prep_input(data, subds = 0, lyrs = NULL, win = NULL)
}
\arguments{
\item{data}{SpatRaster, file path, raster}

\item{subds}{positive integer or character to select a sub-dataset. If zero or "", all sub-datasets are returned (if possible)}

\item{lyrs}{positive integer or character to select a subset of layers (a.k.a. "bands")}

\item{win}{SpatExtent to set a window (area of interest)}
}
\value{
SpatRaster
}
\description{
Prep Incoming Data
}
