% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{single_file_execute}
\alias{single_file_execute}
\title{Internal Exectute Single Summary}
\usage{
single_file_execute(file, w, w_names, name, FUN, rcl, ID)
}
\arguments{
\item{file}{SpatRast object}

\item{w}{weight file}

\item{name}{variable name}

\item{FUN}{summary function}

\item{rcl}{reclassification matrix}

\item{ID}{geometry ID}

\item{w_name}{weight names}
}
\value{
data.table
}
\description{
Internal Exectute Single Summary
}
