% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_grid.R
\name{weighting_grid}
\alias{weighting_grid}
\title{Build Weighting Grid}
\usage{
weighting_grid(file, geom, ID)
}
\arguments{
\item{file}{path to a gridded file (either .tif or .nc)}

\item{geom}{sf object of aggregation units}

\item{ID}{the name of the column providing the unique identified of each geom}
}
\value{
a data.table
}
\description{
Returns a data.table with columns for ID, grid_id, X, Y and weight. By default this object is
sorted on the grid_id
}
